analysis-zero.cfg added for pure KataGo option
analysis-zero.cfg removes human-curated biases that are enabled by default with KataGo. This makes the policy net pure & repeatable, and has a "chance" of making evaluated moves repeatable. It's useful for investigating the engine itself or for analyses that need deterministic policy values. It's not ideal for normal go study, however.