expose k.goban object to scripts
k.goban now available for scripts to play around with. This is an abstract goban of the current position that can export SGF, used for liberty counting, etc.

KataAnswer now generates the k.goban on demand, so the "goban" global has been removed from codeeditor.py.

Also, a brief example of the goban object added to api_demos.

Comments clarified in the Goban class so scripters can get a better feel for it.