fix komi bug
when SGF is imported, komi is set to the value in the SGF. However, this modified the active goban, so that when self.komi was changed later (through the dialog) activeGoban.komi was then out of sync, giving buggy analysis. Therefore, set active goban's komi = self.komi before analysis to keep it consistent. Whew.