WARNING: transition to snake_case
If you have your own KataQuery scripts, please look at the changes made herein!

snake_case will be used for functions and variables inside 'k' and the Goban class, when appropriate. That is, if it doesn't feel like a KataGo data field (which are in mountainCase), snake_case will be used.

The philosophy is that KataAnswer and Goban will be importable (without KataQuery)  later, as a headless analysis library for whoever wants it. snake_case makes much more sense for a library like that.

NOTE: This breaks nearly every script unless changed. Sorry!

KataQuery GUI functions like clearGhosts/clearLog etc. aren't affected, because the gui functions are designed to be easy to type.